+++ src/services/ManifoldKernelDensity.jl
@@ -1,52 +1,50 @@
 
+
 ## ==========================================================================================
 ## helper functions to contruct MKD objects
 ## ==========================================================================================
 
-Ndim(x::ManifoldKernelDensity, w...; kw...) = Ndim(x.belief, w...; kw...)
-Npts(x::ManifoldKernelDensity, w...; kw...) = Npts(x.belief, w...; kw...)
+getPointRepr(x::HomotopyDensity) = eltype(x.data) # TODO use HomotopyDensity{T} style instead
+getManifold(x::HomotopyDensity) = x.manifold
 
-getWeights(x::ManifoldKernelDensity, w...; kw...) = getWeights(x.belief, w...; kw...)
+getPartial(::HomotopyDensity{partial}) where {partial} = partial
+getPartial(x::ManifoldKernelDensity) = getPartial(x.shim)
 
-# getKDERange(x::ManifoldKernelDensity, w...; kw...) = getKDERange(x.belief, w...; kw...)
+Ndim(x::ManifoldKernelDensity, w...; kw...) = Ndim(x.shim, w...; kw...)
+Npts(x::ManifoldKernelDensity, w...; kw...) = Npts(x.shim, w...; kw...)
+
+getWeights(x::ManifoldKernelDensity, w...; kw...) = getWeights(x.shim, w...; kw...)
+
+# getKDERange(x::ManifoldKernelDensity, w...; kw...) = getKDERange(x.shim, w...; kw...)
 # function getKDERange(x::AbstractVector{<:ManifoldKernelDensity}, w...; kw...)
 #     return getKDERange((s -> s.belief).(x), w...; kw...)
 # end
-# getKDEMax(x::ManifoldKernelDensity, w...; kw...) = getKDEMax(x.belief, w...; kw...)
-# getKDEMean(x::ManifoldKernelDensity, w...; kw...) = getKDEMean(x.belief, w...; kw...)
-# getKDEfit(x::ManifoldKernelDensity, w...; kw...) = getKDEfit(x.belief, w...; kw...)
+# getKDEMax(x::ManifoldKernelDensity, w...; kw...) = getKDEMax(x.shim, w...; kw...)
+# getKDEMean(x::ManifoldKernelDensity, w...; kw...) = getKDEMean(x.shim, w...; kw...)
+# getKDEfit(x::ManifoldKernelDensity, w...; kw...) = getKDEfit(x.shim, w...; kw...)
 
-kld(x::ManifoldKernelDensity, w...; kw...) = kld(x.belief, w...; kw...)
-minkld(x::ManifoldKernelDensity, w...; kw...) = minkld(x.belief, w...; kw...)
+kld(x::ManifoldKernelDensity, w...; kw...) = kld(x.shim, w...; kw...)
+minkld(x::ManifoldKernelDensity, w...; kw...) = minkld(x.shim, w...; kw...)
 
-(x::ManifoldKernelDensity)(w...; kw...) = x.belief(w...; kw...)
+(x::ManifoldKernelDensity)(w...; kw...) = x.shim(w...; kw...)
 
-
-function ManifoldKernelDensity(
-    mani::M,
-    bel::B,
-    ::Nothing = nothing,
-    u0::P = zeros(manifold_dimension(mani));
-    infoPerCoord::AbstractVector{<:Real} = ones(getNumberCoords(mani, u0)),
-    partl_cb::Nothing = nothing,
-) where {M <: MB.AbstractManifold, B <: HomotopyDensity, P}
-    return ManifoldKernelDensity{M, B, Nothing, P}(mani, bel, nothing, u0, infoPerCoord)
-end
+getPointRepr(x::ManifoldKernelDensity) = getPointRepr(x.shim)
+getManifold(x::ManifoldKernelDensity) = getManifold(x.shim)
 
 
 function ManifoldKernelDensity(
-    mani::M,
     bel::B,
-    partial_::L,
-    u0::P = zeros(manifold_dimension(mani));
-    infoPerCoord::AbstractVector{<:Real} = ones(getNumberCoords(mani, u0)),
+    partial_::L;
+    # u0::P = zeros(manifold_dimension(mani));
+    infoPerCoord::AbstractVector{<:Real} = ones(getNumberCoords(getManifold(bel), bel.data[1])),
     partl_cb::Union{Nothing, <:Function} = nothing,
-) where {M <: MB.AbstractManifold, B <: HomotopyDensity, L <: AbstractVector{<:Integer}, P}
+) where {B <: HomotopyDensity, L <: Union{<:AbstractVector{<:Integer}, <:Tuple}}
     #
     if isnothing(partl_cb)
-        @warn "WIP partl_cb on MKD constructor helper" maxlog=100
+        @warn "WIP partl_cb on MKD constructor helper" maxlog=10
     end
     partial = _tuple(partial_)
+    mani = getManifold(bel)
     if length(partial) != manifold_dimension(mani)
         # TODO, assuming there are tree and leaf nodes at [1]...
         # @show getKernelTree(bel, 1)
@@ -61,16 +59,17 @@ function ManifoldKernelDensity(
         leaf_kernels_ = view(leaf_kernels, lkm)
         tree_kernels_ .= _intersectpartials.(Ref(mani), view(bel.tree_kernels, tkm), Ref(partial), partl_cb)
         leaf_kernels_ .= _intersectpartials.(Ref(mani), view(bel.leaf_kernels, lkm), Ref(partial), partl_cb)
-        # FIXME update belief to have correct partials
+        # TODO update belief to have correct partials
         bel_ = HomotopyDensity{
             _getprl(eltype(tree_kernels)),
         }(;
-            manifold = bel.manifold,
+            manifold = getManifold(bel),
             data = bel.data,
             weights = bel.weights,
             permute = bel.permute,
             leaf_kernels,
             tree_kernels,
+            infoPerCoord,
             segments = bel.segments,
             _workaround_isdef_leafkernel = bel._workaround_isdef_leafkernel,
             _workaround_isdef_treekernel = bel._workaround_isdef_treekernel,
@@ -78,28 +77,21 @@ function ManifoldKernelDensity(
 
         # call the constructor direct
         # TODO remove _makevec on partials, here and everywhere really.
-        return ManifoldKernelDensity{M, typeof(bel_), L, P}(mani, bel_, _makevec(partial), u0, infoPerCoord)
-        # return ManifoldKernelDensity{M, B, L, P}(mani, bel, partial_, u0, infoPerCoord)
+        return ManifoldKernelDensity{typeof(bel_)}(bel_)
     else
-        # full manifold, therefore equivalent to L::Nothing
-        return ManifoldKernelDensity(mani, bel, nothing, u0; infoPerCoord = infoPerCoord)
+        # full manifold, i.e. partial=nothing
+        return ManifoldKernelDensity(bel)
     end
 end
 
 function ManifoldKernelDensity(
-    mani::M,
     bel::B,
     pl_mask::Union{<:BitVector, <:AbstractVector{<:Bool}},
-    u0::P = zeros(manifold_dimension(mani));
-    infoPerCoord::AbstractVector{<:Real} = ones(getNumberCoords(mani, u0)),
-) where {M <: MB.AbstractManifold, B <: HomotopyDensity, P}
+) where {B <: HomotopyDensity}
     @warn "This constructor is not recommended, as partials have changed somewhat -- possibly erroneous code here..." maxlog=100
     return ManifoldKernelDensity(
-        mani,
         bel,
-        (1:manifold_dimension(mani))[pl_mask],
-        u0;
-        infoPerCoord,
+        # (1:manifold_dimension(getManifold(bel)))[pl_mask], # TODO use tuple instead
     )
 end
 
@@ -156,20 +148,8 @@ function manikde!(
     __partialCovToDefault!(best_cov)
 
     bel = updateBandwidths(mtree, best_cov; partl_cb)
-    infoPerCoord = ones(getNumberCoords(M, pts[1]))
     # return tree with correct bandwidth
-    return ManifoldKernelDensity(M, bel, partial, pts[1], infoPerCoord)
-    # # reuse (heavy lift parts of) earlier tree build
-    # # return manikde!_legacy(M, pts; belmodel = (ignore...) -> updateBandwidths(mtree, best_cov), partial, kw...)
-    # ManifoldKernelDensity(
-    #     M, 
-    #     pts, 
-    #     pts[1]; 
-    #     belmodel = (ignore...) -> updateBandwidths(mtree, best_cov; partl_cb), 
-    #     partial, 
-    #     partl_cb,
-    #     kw...
-    # )
+    return ManifoldKernelDensity(bel)
 end
 
 ## ==========================================================================================
@@ -200,45 +180,41 @@ function _partialCovToDefault!(p::Union{<:Tuple, <:AbstractVector{<:Integer}}, m
     return m
 end
 
-function _getFieldPartials(
-    mkd::ManifoldKernelDensity{M, B, Nothing},
-    field::Function,
-    aspartial::Bool = true,
-) where {M, B}
-    return field(mkd)
-end
 
 function _getFieldPartials(
-    mkd::ManifoldKernelDensity{M, B, <:AbstractVector},
+    mkd::HomotopyDensity{partial},
     field::Function,
     aspartial::Bool = true,
-) where {M, B}
+) where {partial}
+    if isnothing(partial)
+        return field(mkd)
+    end
     _length(x::AbstractMatrix) = length(diag(x))
     _length(x::AbstractVector) = length(x)
     val = field(mkd)
-    if aspartial && (_length(val) == length(mkd._partial))
+    if aspartial && (_length(val) == length(getPartial(mkd)))
         return val
-    elseif !aspartial && (_length(val) == length(mkd._partial))
-        val_ = zeros(manifold_dimension(mkd.manifold))
-        val_[mkd._partial] .= val
+    elseif !aspartial && (_length(val) == length(getPartial(mkd)))
+        val_ = zeros(manifold_dimension(getManifold(mkd)))
+        val_[getPartial(mkd)] .= val
         return val_
-    elseif aspartial && (_length(val) == manifold_dimension(mkd.manifold))
-        return val[mkd._partial]
-    elseif !aspartial && (_length(val) == manifold_dimension(mkd.manifold))
+    elseif aspartial && (_length(val) == manifold_dimension(getManifold(mkd)))
+        return val[getPartial(mkd)]
+    elseif !aspartial && (_length(val) == manifold_dimension(getManifold(mkd)))
         return val
     else
         error(
-            "unknown size MKD.$(field) with partial length=$(length(mkd._partial)) vs length=$(_length(val)) --- and value=$val",
+            "unknown size MKD.$(field) with partial length=$(length(getPartial(mkd))) vs length=$(_length(val)) --- and value=$val",
         )
     end
 end
 
 function getInfoPerCoord(mkd::ManifoldKernelDensity, aspartial::Bool = true)
-    return _getFieldPartials(mkd, x -> x.infoPerCoord, aspartial)
+    return _getFieldPartials(mkd.shim, x -> x.infoPerCoord, aspartial)
 end
 
 function getBandwidth(mkd::ManifoldKernelDensity, aspartial::Bool = true)
-    return _getFieldPartials(mkd, x -> getBW(x)[1], aspartial)
+    return _getFieldPartials(mkd.shim, x -> getBW(x)[1], aspartial)
 end
 
 
@@ -247,8 +223,9 @@ end
 
 Return true if this ManifoldKernelDensity is a partial.
 """
-isPartial(mkd::ManifoldKernelDensity{M, B, L}) where {M, B, L} = true
-isPartial(mkd::ManifoldKernelDensity{M, B, Nothing}) where {M, B} = false
+isPartial(hd::HomotopyDensity) = !isnothing(getPartial(hd))
+isPartial(mkd::ManifoldKernelDensity) = isPartial(mkd.shim)
+
 
 """
     $SIGNATURES
@@ -263,63 +240,49 @@ DevNotes
 - Currently converts down to manifold from matrix of coordinates (legacy), to be deprecated TODO
 """
 function getPoints(
-    x::ManifoldKernelDensity{M, B},
-    ::Bool = true; # aspartial unused
-    permute::Bool = true,
-) where {M <: AbstractManifold, B}
-    return getPoints(x.belief; permute)
-    # return _matrixCoordsToPoints(x.manifold, getPoints(x.belief), x._u0)
-end
-
-
-function getPoints(
-    x::ManifoldKernelDensity{M, B, L},
+    x::ManifoldKernelDensity,
     aspartial::Bool = true;
     permute::Bool = true,
-) where {M <: AbstractManifold, B <: HomotopyDensity, L <: AbstractVector{Int}}
+)
     #
-    pts = getPoints(x.belief; permute)
+    pts = getPoints(x.shim; permute)
 
-    if (L === nothing) && !aspartial
+    if !isPartial(x) && !aspartial
         error("MKD getPoints aspartial=true but MKD is not partial")
         return pts
     end
 
-    Mp, Rp, lkup = getManifoldPartial(x.manifold, x._partial, x._u0)
+    Mp, Rp, lkup = getManifoldPartial(getManifold(x), getPartial(x), pts[1])
 
     vecP = Vector{typeof(Rp)}(undef, length(pts))
     for (j,pt) in enumerate(pts)
         vecP[j] =  lkup(pt)
     end
     return vecP
-
-    # (x.manifold, x._u0)
-    # x._partial
-    # return _matrixCoordsToPoints(M_, pts_, u0_)
 end
 
 
 function getBW(
-    x::ManifoldKernelDensity{M, B, L},
+    x::ManifoldKernelDensity{B},
     asPartial::Bool = true;
     kw...,
-) where {M, B, L}
-    bws = getBW(x.belief; kw...)
-    if L !== Nothing && asPartial
-        return (bw->view(bw, x._partial)).(bws)
+) where {B}
+    bws = getBW(x.shim; kw...)
+    if isPartial(x) && asPartial
+        return (bw->view(bw, getPartial(x))).(bws)
     end
     return bws
 end
 
 # TODO check that partials / marginals are sampled correctly
-function sample(x::ManifoldKernelDensity{M, B, L, P}, N::Integer = 1) where {M, B, L, P}
+function sample(x::ManifoldKernelDensity, N::Integer = 1)
     # get legacy matrix of coordinates and selected labels
-    coords, lbls = sample(x.belief, N)
-
+    belief = x.shim
+    coords, lbls = sample(belief, N)
     # pack samples into vector of point type P
-    vecP = Vector{P}(undef, N)
+    vecP = Vector{eltype(belief.data)}(undef, N)
     for j = 1:N
-        vecP[j] = makePointFromCoords(x.manifold, view(coords, :, j), x._u0)
+        vecP[j] = makePointFromCoords(getManifold(x), view(coords, :, j), belief.data[1])
     end
 
     return vecP, lbls
@@ -332,8 +295,8 @@ Distributions.variate_form(mkd::ManifoldKernelDensity) = Ndim(mkd) == 1 ? Univar
 function Random.rand(mkd::ManifoldKernelDensity)
     return _rand(Distributions.variate_form(mkd), mkd)
 end
-_rand(::Type{Univariate}, mkd::ManifoldKernelDensity) = sample(mkd.belief, 1)[1][:][]
-_rand(::Type{Multivariate}, mkd::ManifoldKernelDensity) = sample(mkd.belief, 1)[1][:]
+_rand(::Type{Univariate}, mkd::ManifoldKernelDensity) = sample(mkd.shim, 1)[1][:][]
+_rand(::Type{Multivariate}, mkd::ManifoldKernelDensity) = sample(mkd.shim, 1)[1][:]
 
 function resample(x::ManifoldKernelDensity, N::Int)
     pts = if N < Npts(x)
@@ -344,45 +307,42 @@ function resample(x::ManifoldKernelDensity, N::Int)
         _pts
     end
     return ManifoldKernelDensity(
-        x.manifold,
-        pts,
-        x._u0;
-        partial = x._partial,
-        infoPerCoord = x.infoPerCoord,
+        getManifold(x),
+        pts;
+        partial = getPartial(x),
+        infoPerCoord = x.shim.infoPerCoord,
     )
 end
 
-function Base.show(io::IO, mkd::ManifoldKernelDensity{M, B, L, P}) where {M, B, L, P}
+function Base.show(io::IO, mkd::ManifoldKernelDensity{B}) where {B}
     _round(s::AbstractArray; kw...) = round.(s[:]; kw...)
     _round(s::AbstractVector{<:AbstractMatrix}; kw...) = round.(s[1][:]; kw...)
 
     printstyled(io, "ManifoldKernelDensity{"; bold = true, color = :blue)
     println(io)
-    printstyled(io, "    M"; bold = true, color = :magenta)
-    print(io, " = ", M, ",")
-    println(io)
+    # FIXME restore after HomotopyDensity refactor
+    # printstyled(io, "    M"; bold = true, color = :magenta)
+    # print(io, " = ", M, ",")
+    # println(io)
     printstyled(io, "    B"; bold = true, color = :magenta)
     print(io, " = ", B, ",")
     println(io)
-    printstyled(io, "    L"; bold = true, color = :magenta)
-    print(io, " = ", L, ",")
-    println(io)
-    printstyled(io, "    P"; bold = true, color = :magenta)
-    print(io, " = ", P)
-    println(io)
+    # FIXME restore after HomotopyDensity refactor
+    # printstyled(io, "    L"; bold = true, color = :magenta)
+    # print(io, " = ", L, ",")
+    # println(io)
     println(io, " }(")
-    println(io, "  Npts:  ", Npts(mkd.belief))
-    print(io, "  dims:  ", Ndim(mkd.belief))
-    printstyled(io, isPartial(mkd) ? "* --> $(length(mkd._partial))" : ""; bold = true)
+    println(io, "  Npts:  ", Npts(mkd.shim))
+    print(io, "  dims:  ", Ndim(mkd.shim))
+    printstyled(io, isPartial(mkd) ? "* --> $(length(getPartial(mkd)))" : ""; bold = true)
     println(io)
-    println(io, "  prtl:   ", mkd._partial)
-    bw = (getBW(mkd.belief).^2)[:, 1]
-    pvec = isPartial(mkd) ? mkd._partial : collect(1:length(bw))
+    println(io, "  prtl:   ", getPartial(mkd))
+    bw = (getBW(mkd.shim).^2)[:, 1]
+    pvec = isPartial(mkd) ? getPartial(mkd) : collect(1:length(bw))
     println(io, "  bws:   ", getBandwidth(mkd, true) |> x -> _round(x; digits = 4)) # .|> x->round(x,digits=4))
     println(io, "  ipc:   ", getInfoPerCoord(mkd, true) .|> x -> round(x; digits = 4))
     print(io, "   mean: ")
     try
-        # mn = mean(mkd.manifold, getPoints(mkd, false))
         mn = mean(mkd)
         if mn isa ProductRepr # TODO UPDATE to ArrayPartition only, discontinued use of ProductRepr long ago.
             println(io)
@@ -406,12 +366,12 @@ end
 
 # override
 function marginal(
-    x::ManifoldKernelDensity{M, B},
+    x::ManifoldKernelDensity{B},
     dims::AbstractVector{<:Integer},
-) where {M <: AbstractManifold, B}
+) where {B}
     #
-    ldims::Vector{Int} = _makevec(_intersect(x._partial, collect(dims)))
-    return ManifoldKernelDensity(x.manifold, x.belief, ldims, x._u0)
+    ldims::Vector{Int} = _makevec(_intersect(getPartial(x), collect(dims)))
+    return ManifoldKernelDensity(x.shim, ldims)
 end
 
 
@@ -442,7 +402,7 @@ function antimarginal(
     # new coord partials must be placed into a full dimension point, thats why we use u0
     nPts = Vector{typeof(u0)}(undef, length(pts))
     for i in eachindex(pts)
-        setPointPartial!(newM, nPts, mkd.manifold, pts, newpartial, i)
+        setPointPartial!(newM, nPts, getManifold(mkd), pts, newpartial, i)
     end
 
     # also update metadata elements